// jobhunter-FE/src/types/backend.d.ts
    export interface IBackendRes<T> {
        error?: string | string[];
        message: string;
        statusCode: number | string;
        data?: T;
    }

    export interface IModelPaginate<T> {
        meta: {
            page: number;
            pageSize: number;
            pages: number;
            total: number;
        },
        result: T[]
    }
// ...
    export interface ISkill {
        id?: string;
        name?: string;
        createdBy?: string;
        isDeleted?: boolean;
        deletedAt?: boolean | null;
        createdAt?: string;
        updatedAt?: string;
    }
// ...
    export interface IJob {
        id?: string;
        name: string;
        skills: ISkill[];
        company?: {
            id: string;
            name: string;
            logo?: string;
        }
        location: string;
        salary: number;
        quantity: number;
        level: string;
        description: string;
        startDate: Date;
        endDate: Date;
        active: boolean;

        createdBy?: string;
        isDeleted?: boolean;
        deletedAt?: boolean | null;
        createdAt?: string;
        updatedAt?: string;
    }